+++ tests/test_providers/patterns.py
@@ -79,6 +79,8 @@ POSTAL_CODE_REGEX = {
 HOME_PAGE = r'http[s]?://(?:[a-zA-Z]|[0-9]|[$_@.&+-]' \
             r'|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+'
 
+HOST_NAME = r'[a-zA-Z0-9-]+(\.[a-zA-Z0-9-]+)*'
+
 UUID_REGEX = r'[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}'
 
 HEX_COLOR = r'#(?:[a-fA-F0-9]{3}|[a-fA-F0-9]{6})\b'